refactor: corrigir caminho de importação para o componente MyMealsPage
