+++ src/routes/index.tsx
@@ -14,7 +14,7 @@ import { AddFoodPage } from '@/pages/Foods/AddFood/add-food-view-model'
 import { EditFoodPage } from '@/pages/Foods/EditFood/edit-food-view-model'
 import { FoodDetailsPage } from '@/pages/Foods/FoodDetails/food-details-view-model'
 
-import { MyMealsPage } from '@/pages/Meals/MyMeals/page'
+import { MyMealsPage } from '@/pages/Meals/my-meals/my-meals-view-model'
 import { CreateMealPage } from '@/pages/Meals/create-meal/create-meal-view-model'
 import { EditMealPage } from '@/pages/Meals/EditMeal/page'
